CommunicationClient Properties and Factory extension points as events
Contains new features that enables adding metadata to the communication client through a new Properties dictionary on the client.
Also adds two new extension points in the RestCommunicationClientFactory exposed as events. The events are named ClientCreated and ClientValidating indicating when they are fired.

This allows subscribers to store metadata in client, which could be done at creation time through the ClientCreated event, and can be used to determine validity of a client by examining the client properties and setting the IsValid property on the event args of the ClientValidating event.